requirements.txt
fastapi
uvicorn
pyjwt
cryptography